// src/app/models/client-homeless-histories.ts
export class ClientHomelessHistory {
    id: number = undefined;
    client_id: number = undefined;
    first_time_homeless: boolean = undefined;
    date_became_homeless: Date = undefined;
    reason_for_homelessness: string = undefined;
    note: string = undefined;
    created_at: Date = undefined;
    updated_at: Date = undefined;
  }
  